+++ CHANGELOG.md
@@ -66,6 +66,8 @@ On the other hand, this also allowed us to deal with further issues that contain
   - Moved `get_splits()` method from `small_text.utils.classification` to `small_text.data.splits`.
 - Dependencies
   - Raised minimum required setfit version to 1.1.0.
+- Active Learner
+  - Revised serialization interface: `save()` and `load()` now operate on folders.
 - Classification:
   - The `initialize()` methods of all PyTorch-classifiers (KimCNN, TransformerBasedClassification, SetFitClassification) are now more unified. ([#57](https://github.com/webis-de/small-text/issues/57))
   - KimCNNClassifier / TransformerBasedClassification: model selection is now disabled by default. Also, it no longer saves models when disabled, thereby greatly reducing the runtime.